requirements-dev.txt
Flask
Flask-Login
Flask-OpenID
Flask-SQLAlchemy
Flask-WTF
Jinja2
MarkupSafe
SQLAlchemy
WTForms
Werkzeug
itsdangerous
oauthlib
python-social-auth
python3-openid
requests
requests-oauthlib
six
psycopg2
gunicorn
